pyproject.toml
[tool.poetry]
name = "Chatbot"
version = "0.1.0"
description = "Local Ollama PDF RAG Chatbot"
readme = "README.md"
license = "WTFPL"
authors = ["Catik mccatik@gmail.com"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

# Default PyPI repository.
aiofiles = "24.1.0"
aiohappyeyeballs = "2.4.6"
aiohttp = "3.11.13"
aiosignal = "1.3.2"
altair = "5.5.0"
annotated-types = "0.7.0"
antlr4-python3-runtime = "4.9.3"
anyio = "4.8.0"
asgiref = "3.8.1"
async-timeout = "4.0.3"
attrs = "25.1.0"
babel = "2.17.0"
backoff = "2.2.1"
backrefs = "5.8"
bcrypt = "4.3.0"
beautifulsoup4 = "4.13.3"
blinker = "1.9.0"
build = "1.2.2.post1"
cachetools = "5.5.2"
certifi = "2025.1.31"
cffi = "1.17.1"
chardet = "5.2.0"
charset-normalizer = "3.4.1"
chroma-hnswlib = "0.7.6"
chromadb = "0.6.3"
click = "8.1.8"
colorama = "0.4.6"
coloredlogs = "15.0.1"
contourpy = "1.3.1"
coverage = "7.4.0"
cryptography = "44.0.1"
cycler = "0.12.1"
dataclasses-json = "0.6.7"
Deprecated = "1.2.18"
distro = "1.9.0"
durationpy = "0.9"
effdet = "0.4.1"
emoji = "2.14.1"
et_xmlfile = "2.0.0"
eval_type_backport = "0.2.2"
exceptiongroup = "1.2.2"
fastapi = "0.115.10"
filelock = "3.17.0"
filetype = "1.2.0"
Flask = "3.1.0"
flatbuffers = "25.2.10"
fonttools = "4.56.0"
frozenlist = "1.5.0"
fsspec = "2025.2.0"
ghp-import = "2.1.0"
gitdb = "4.0.12"
GitPython = "3.1.44"
google-api-core = "2.24.1"
google-auth = "2.38.0"
google-cloud-vision = "3.10.0"
googleapis-common-protos = "1.68.0"
greenlet = "3.1.1"
griffe = "1.6.0"
grpcio = "1.70.0"
grpcio-status = "1.70.0"
h11 = "0.14.0"
htbuilder = "0.9.0"
html5lib = "1.1"
httpcore = "1.0.7"
httptools = "0.6.4"
httpx = "0.27.2"
httpx-sse = "0.4.0"
huggingface-hub = "0.29.1"
humanfriendly = "10.0"
idna = "3.10"
importlib_metadata = "8.5.0"
importlib_resources = "6.5.2"
iniconfig = "2.0.0"
iopath = "0.1.10"
itsdangerous = "2.2.0"
Jinja2 = "3.1.5"
joblib = "1.4.2"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
kiwisolver = "1.4.8"
kubernetes = "32.0.1"
langchain = "0.3.14"
langchain-community = "0.3.14"
langchain-core = "0.3.29"
langchain-ollama = "0.2.2"
langchain-text-splitters = "0.3.5"
langdetect = "1.0.9"
langsmith = "0.2.11"
layoutparser = "0.3.4"
lxml = "5.3.1"
Markdown = "3.7"
markdown-it-py = "3.0.0"
MarkupSafe = "3.0.2"
marshmallow = "3.26.1"
matplotlib = "3.10.1"
mdurl = "0.1.2"
mergedeep = "1.3.4"
mkdocs = "1.6.1"
mkdocs-autorefs = "1.4.0"
mkdocs-get-deps = "0.2.0"
mkdocs-material = "9.6.6"
mkdocs-material-extensions = "1.3.1"
mkdocstrings = "0.28.2"
mkdocstrings-python = "1.16.2"
mmh3 = "5.1.0"
monotonic = "1.6"
mpmath = "1.3.0"
multidict = "6.1.0"
mypy-extensions = "1.0.0"
narwhals = "1.28.0"
ndjson = "0.3.1"
nest-asyncio = "1.6.0"
networkx = "3.4.2"
nltk = "3.9.1"
numpy = "1.26.4"
oauthlib = "3.2.2"
olefile = "0.47"
ollama = "0.4.4"
omegaconf = "2.3.0"
onnx = "1.17.0"
onnxruntime = "1.20.1"
onnxruntime-gpu = "1.20.1"
opencv-python = "4.11.0.86"
openpyxl = "3.1.5"
opentelemetry-api = "1.30.0"
opentelemetry-exporter-otlp-proto-common = "1.30.0"
opentelemetry-exporter-otlp-proto-grpc = "1.30.0"
opentelemetry-instrumentation = "0.51b0"
opentelemetry-instrumentation-asgi = "0.51b0"
opentelemetry-instrumentation-fastapi = "0.51b0"
opentelemetry-proto = "1.30.0"
opentelemetry-sdk = "1.30.0"
opentelemetry-semantic-conventions = "0.51b0"
opentelemetry-util-http = "0.51b0"
orjson = "3.10.15"
overrides = "7.7.0"
packaging = "24.2"
paginate = "0.5.7"
pandas = "2.2.3"
pathspec = "0.12.1"
pdf2image = "1.17.0"
pdfminer-six = "20231228"
pdfplumber = "0.11.4"
pi_heif = "0.21.0"
pikepdf = "9.5.2"
pillow = "10.4.0"
platformdirs = "4.3.6"
pluggy = "1.5.0"
portalocker = "3.1.1"
posthog = "3.17.0"
propcache = "0.3.0"
proto-plus = "1.26.0"
protobuf = "5.29.2"
psutil = "7.0.0"
pyarrow = "19.0.1"
pyasn1 = "0.6.1"
pyasn1_modules = "0.4.1"
pycocotools = "2.0.8"
pycparser = "2.22"
pydantic = "2.10.4"
pydantic-settings = "2.8.1"
pydantic_core = "2.27.2"
pydeck = "0.9.1"
Pygments = "2.19.1"
pymdown-extensions = "10.14.3"
pypandoc = "1.15"
pyparsing = "3.2.1"
pypdf = "5.3.0"
pypdfium2 = "4.30.1"
PyPika = "0.48.9"
pyproject_hooks = "1.2.0"
pyreadline3 = "3.5.4"
pytest = "7.4.4"
pytest-cov = "4.1.0"
python-dateutil = "2.9.0.post0"
python-docx = "1.1.2"
python-dotenv = "1.0.1"
python-iso639 = "2025.2.18"
python-magic = "0.4.27"
python-multipart = "0.0.20"
python-oxmsg = "0.0.2"
python-pptx = "1.0.2"
pytz = "2025.1"
pywin32 = "308"
PyYAML = "6.0.2"
pyyaml_env_tag = "0.1"
RapidFuzz = "3.12.1"
referencing = "0.36.2"
regex = "2024.11.6"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
requests-toolbelt = "1.0.0"
rich = "13.9.4"
rpds-py = "0.23.1"
rsa = "4.9"
safetensors = "0.5.3"
scipy = "1.15.2"
shellingham = "1.5.4"
six = "1.17.0"
smmap = "5.0.2"
sniffio = "1.3.1"
soupsieve = "2.6"
SQLAlchemy = "2.0.38"
starlette = "0.46.0"
streamlit = "1.40.0"
sympy = "1.13.1"
tenacity = "9.0.0"
timm = "1.0.15"
tokenizers = "0.21.0"
toml = "0.10.2"
tomli = "2.2.1"
tornado = "6.4.2"
tqdm = "4.67.1"
transformers = "4.49.0"
typer = "0.15.2"
typing-inspect = "0.9.0"
typing_extensions = "4.12.2"
tzdata = "2025.1"
unstructured = "0.16.14"
unstructured-client = "0.30.6"
unstructured-inference = "0.8.1"
unstructured-pytesseract = "0.3.13"
urllib3 = "2.3.0"
uvicorn = "0.34.0"
watchdog = "5.0.3"
watchfiles = "1.0.4"
webencodings = "0.5.1"
websocket-client = "1.8.0"
websockets = "15.0"
Werkzeug = "3.1.3"
wrapt = "1.17.2"
xlrd = "2.0.1"
XlsxWriter = "3.2.2"
yarl = "1.18.3"
zipp = "3.21.0"

# PyTorch-specific dependencies (explicitly fetched from the custom repository)
torch = { version = "2.6.0", source = "pytorch" }
torchvision = { version = "0.21.0", source = "pytorch" }
colorlog = "^6.9.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
